[MIG] pos_pricelist_customer_required: migration 18.0
