+++ pos_pricelist_customer_required/models/product_pricelist.py
@@ -9,3 +9,8 @@ class ProductPricelist(models.Model):
     _inherit = "product.pricelist"
 
     pos_require_customer = fields.Boolean(default=True)
+
+    def _load_pos_data_fields(self, config_id):
+        res = super()._load_pos_data_fields(config_id)
+        res.append("pos_require_customer")
+        return res
