--- pos_pricelist_customer_required/templates/assets.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template id="assets" inherit_id="point_of_sale.assets">
-        <xpath expr=".">
-            <script
-                type="text/javascript"
-                src="/pos_require_customer/static/src/js/models.js"
-            />
-            <script
-                type="text/javascript"
-                src="/pos_require_customer/static/src/js/PaymentScreen.js"
-            />
-        </xpath>
-    </template>
-</odoo>
